refactor: improve components
